requirements.txt
azure-iot-device~=2.12.0
cryptography~=41.0.3
